Move sample artifact download endpoint to `caches.api`
